requirements.txt
anthropic>=0.49.0,<1.0.0
openai>=1.64.0,<2.0.0
pandas>=2.2.0,<3.0.0
scikit-learn>=1.5.0,<2.0.0
matplotlib>=3.9.0,<4.0.0
seaborn>=0.13.0,<0.14.0
python-dotenv>=1.0.0,<2.0.0
tqdm>=4.67.0,<5.0.0
